+++ src/userobject.cpp
@@ -41,29 +41,13 @@ const UserObject UserObject::nothing;
 UserObject::UserObject()
     : m_class(nullptr)
     , m_holder()
-    , m_parent()
-    , m_child(nullptr)
 {
-}
-
-UserObject::UserObject(const UserObject& parent, const UserProperty& member)
-    : m_class(&member.getClass())
-    , m_holder()
-    , m_parent(new ParentObject(parent, member))
-    , m_child(nullptr)
-{
-    m_parent->object.m_child = this;
 }
 
 UserObject::UserObject(const UserObject& copy)
     : m_class(copy.m_class)
     , m_holder(copy.m_holder)
-    , m_parent(copy.m_parent
-               ? new ParentObject(copy.m_parent->object, copy.m_parent->member)
-               : nullptr)
 {
-    if (m_parent)
-        m_parent->object.m_child = this;
 }
 
 void* UserObject::pointer() const
@@ -71,13 +55,6 @@ void* UserObject::pointer() const
     if (m_holder)
     {
         return m_holder->object();
-    }
-    else if (m_parent)
-    {
-        // warning: this may not be multi-thread safe,
-        // but it is required to make the returned value persistent
-        m_parent->lastValue = m_parent->member.get(m_parent->object).to<UserObject>();
-        return m_parent->lastValue.pointer();
     }
     else
     {
@@ -124,18 +101,8 @@ Value UserObject::call(IdRef function, const Args& args) const
 
 UserObject& UserObject::operator=(const UserObject& other)
 {
-    std::unique_ptr<ParentObject>
-        parent(other.m_parent
-               ? new ParentObject(other.m_parent->object, other.m_parent->member)
-               : nullptr);
-
     m_class = other.m_class;
     m_holder = other.m_holder;
-    m_parent.swap(parent);
-    m_child = nullptr;
-
-    if (m_parent)
-        m_parent->object.m_child = this;
 
     return *this;
 }
@@ -145,11 +112,6 @@ bool UserObject::operator == (const UserObject& other) const
     if (m_holder && other.m_holder)
     {
         return m_holder->object() == other.m_holder->object();
-    }
-    else if (m_parent && other.m_parent)
-    {
-        return (&m_parent->member == &other.m_parent->member)
-            && (m_parent->object == other.m_parent->object);
     }
     else if (!m_class && !other.m_class)
     {
@@ -167,29 +129,9 @@ bool UserObject::operator < (const UserObject& other) const
         {
             return m_holder->object() < other.m_holder->object();
         }
-        else
-        {
-            return other.m_parent != nullptr;
-        }
     }
-    else if (m_parent)
-    {
-        if (other.m_parent)
-        {
-            if (&m_parent->member != &other.m_parent->member)
-                return &m_parent->member < &other.m_parent->member;
-            else
-                return m_parent->object < other.m_parent->object;
-        }
-        else
-        {
-            return false;
-        }
-    }
-    else
-    {
-        return other.m_holder || other.m_parent;
-    }
+    assert(0);
+    return false;
 }
 
 void UserObject::set(const Property& property, const Value& value) const
@@ -199,54 +141,11 @@ void UserObject::set(const Property& property, const Value& value) const
         // Just forward to the property, no extra processing required
         property.setValue(*this, value);
     }
-    else if (m_parent)
-    {
-        // Find the root object
-        const UserObject* root = this;
-        while (root->m_parent->object.m_parent)
-            root = &root->m_parent->object;
-
-        // Launch the recursive process
-        root->cascadeSet(root->m_parent->object, property, value);
-    }
     else
     {
         // Error, null object
         PONDER_ERROR(NullObject(m_class));
     }
-}
-
-void UserObject::cascadeSet(const UserObject& object,
-                            const Property& property,
-                            const Value& value) const
-{
-    // @todo Manually check the access (read / write) to the properties,
-    // as we bypass the standard path?
-
-    // Get the temporary object to modify
-    UserObject sub = m_parent->member.getValue(object).to<UserObject>();
-
-    // Make sure that the underlying object is *really* modifiable.
-    // If m_holder was holding a const reference to an object, this will
-    // replace it with a holder storing a copy of the object.
-    detail::AbstractObjectHolder* holder = sub.m_holder->getWritable();
-    bool same = (holder == sub.m_holder.get());
-    if (!same)
-        sub.m_holder.reset(holder);
-
-    if (m_child)
-    {
-        // ...continue to recurse if it's not the actual property to modify
-        m_child->cascadeSet(sub, property, value);
-    }
-    else
-    {
-        // ...apply the value if it's the actual property to modify
-        property.setValue(sub, value);
-    }
-
-    // Assign the modified object back to its parent
-    m_parent->member.setValue(object, sub);
 }
 
 } // namespace ponder