+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Ponder Changelog
 ### 1.4
 
 - Added support for binding to static functions (issue #46).
+- Removed UserObject parent-child. Simplification, unused.
 - Class::create() allows object creation without an Args list.
 - Class::tryFunction() and Class::tryProperty() added.